mv internal dir, mv test, add comments
