+++ tests/tt_metal/distributed/test_dispatch_context.cpp
@@ -4,13 +4,10 @@
 
 #include <gtest/gtest.h>
 
-#include <chrono>
 #include <cstdint>
 #include <memory>
 #include <numeric>
 #include <set>
-#include <thread>
-#include <unordered_set>
 #include <utility>
 #include <vector>
 
@@ -19,7 +16,6 @@
 #include <tt-metalium/kernel_types.hpp>
 #include <tt-metalium/distributed.hpp>
 #include <tt-metalium/experimental/dispatch_context.hpp>
-#include <tt-metalium/internal/service/service_core_manager.hpp>
 #include <tt-metalium/mesh_buffer.hpp>
 #include <tt-metalium/mesh_coord.hpp>
 #include <tt-metalium/mesh_device.hpp>
@@ -28,9 +24,7 @@
 #include <tt-metalium/tt_metal.hpp>
 
 #include "impl/context/metal_context.hpp"
-#include "llrt/llrt.hpp"
 #include "llrt/tt_cluster.hpp"
-#include "llrt/hal/generated/dev_msgs.hpp"
 #include "tests/tt_metal/distributed/utils.hpp"
 #include <umd/device/types/arch.hpp>
 
@@ -285,418 +279,6 @@ TEST_F(DispatchContextFixture, AsyncSdStatePreservedAcrossFdTransition) {
 
     // Verify async SD state survived the round-trip
     EXPECT_TRUE(experimental::DispatchContext::get().is_asynchronous_slow_dispatch_enabled(mesh_device_.get()));
-}
-
-// Verify that claiming all free FD dispatch cores as service cores:
-//   (a) caps compute_with_storage_grid_size() to the FD grid in SD mode (conservative cap),
-//   (b) leaves no idle dispatch cores when FD is re-initialized.
-TEST_F(DispatchContextFixture, ServiceCoreGridCap) {
-    const auto& rt_options = MetalContext::instance().rtoptions();
-    if (rt_options.get_fast_dispatch()) {
-        GTEST_SKIP() << "This test requires Slow Dispatch mode.";
-    }
-    const auto& cluster = MetalContext::instance().get_cluster();
-    if (!cluster.is_ubb_galaxy() && cluster.arch() != tt::ARCH::BLACKHOLE) {
-        GTEST_SKIP() << "Service-core APIs require BH or UBB Galaxy.";
-    }
-
-    auto mesh_device = MeshDevice::create(MeshDeviceConfig(MeshShape(1, 1)));
-    IDevice* device = mesh_device->get_device(MeshCoordinate(0, 0));
-
-    // Phase 1: record full SD grid (no services active).
-    const CoreCoord sd_grid = device->compute_with_storage_grid_size();
-
-    // Phase 2: init FD, record FD grid, claim all free dispatch cores.
-    experimental::DispatchContext::get().initialize_fast_dispatch(mesh_device.get());
-    const CoreCoord fd_grid = device->compute_with_storage_grid_size();
-    ASSERT_LT(fd_grid.x * fd_grid.y, sd_grid.x * sd_grid.y)
-        << "FD grid must be smaller than SD grid (dispatch column excluded by YAML)";
-
-    const auto free_cores = internal::ServiceCoreManager::get().get_claimable_cores(device);
-
-    internal::ServiceCoreManager::get().claim(device, free_cores);
-
-    // Attempting to claim the same cores again must fail — claim guard is independent
-    // of the pool filter and catches double-claim regardless of FD state.
-    EXPECT_THROW(internal::ServiceCoreManager::get().claim(device, free_cores), std::exception);
-
-    // Phase 3: terminate + re-init FD — pool filter must exclude all claimed cores,
-    // leaving no idle dispatch cores for new FD kernels.
-    experimental::DispatchContext::get().terminate_fast_dispatch(mesh_device.get());
-    experimental::DispatchContext::get().initialize_fast_dispatch(mesh_device.get());
-
-    // All cores claimed — get_claimable_cores() must fatal rather than return an empty vector.
-    EXPECT_THROW(internal::ServiceCoreManager::get().get_claimable_cores(device), std::exception);
-
-    experimental::DispatchContext::get().terminate_fast_dispatch(mesh_device.get());
-
-    // Phase 4: back in SD — grid must be capped to the FD grid (conservative cap).
-    const CoreCoord capped_grid = device->compute_with_storage_grid_size();
-    EXPECT_EQ(capped_grid.x, fd_grid.x) << "SD grid x must be capped to FD grid x when service cores are claimed";
-    EXPECT_EQ(capped_grid.y, fd_grid.y) << "SD grid y must be capped to FD grid y when service cores are claimed";
-
-    // Phase 5: release service cores and verify the full SD grid is restored.
-    internal::ServiceCoreManager::get().release(device, free_cores);
-
-    const CoreCoord restored_grid = device->compute_with_storage_grid_size();
-    EXPECT_EQ(restored_grid.x, sd_grid.x) << "SD grid x must be restored to original after release";
-    EXPECT_EQ(restored_grid.y, sd_grid.y) << "SD grid y must be restored to original after release";
-
-    internal::ServiceCoreManager::get().on_device_close(device->id());
-}
-
-// Persistent service pipecleaner: test a Persistent kernel on FD cores (survive multiple FD init + teardown cycles)
-// Kernel increments a counter in L1 until a stop flag is set
-// We verify the counter is still incrementing after each FD lifecycle transition
-TEST_F(DispatchContextFixture, PersistentServiceMultiCycle) {
-    const auto& rt_options = MetalContext::instance().rtoptions();
-    if (rt_options.get_fast_dispatch()) {
-        GTEST_SKIP() << "This test requires Slow Dispatch mode.";
-    }
-    const auto& cluster = MetalContext::instance().get_cluster();
-    if (!cluster.is_ubb_galaxy() && cluster.arch() != tt::ARCH::BLACKHOLE) {
-        GTEST_SKIP() << "Service-core APIs require BH or UBB Galaxy.";
-    }
-
-    auto mesh_device = MeshDevice::create(MeshDeviceConfig(MeshShape(1, 1)));
-    IDevice* device = mesh_device->get_device(MeshCoordinate(0, 0));
-
-    uint32_t single_tile_size = ::tt::tile_size(DataFormat::UInt32);
-    const uint32_t num_tiles = 64;
-    const uint32_t num_programs = 1;
-
-    CoreRangeSet shard_grid(CoreRange({0, 0}, {1, 1}));
-    const uint32_t num_cores = 4;
-    const uint32_t tiles_per_shard = num_tiles / num_cores;
-    std::array<uint32_t, 2> shard_shape = {tiles_per_shard * single_tile_size, 1};
-    std::array<uint32_t, 2> page_shape = {single_tile_size, 1};
-    std::array<uint32_t, 2> tensor2d_shape = {num_tiles, 1};
-    ShardSpecBuffer shard_spec(shard_grid, shard_shape, ShardOrientation::ROW_MAJOR, page_shape, tensor2d_shape);
-
-    DeviceLocalBufferConfig fd_l1_config{
-        .page_size = single_tile_size,
-        .buffer_type = BufferType::L1,
-        .sharding_args = BufferShardingArgs(shard_spec, TensorMemoryLayout::HEIGHT_SHARDED),
-        .bottom_up = false};
-    ReplicatedBufferConfig fd_l1_global{.size = num_tiles * single_tile_size};
-
-    constexpr int num_cycles = 5;
-    for (int cycle = 0; cycle < num_cycles; ++cycle) {
-        // Phase 1: FD up, claim a free core, drop FD.
-        experimental::DispatchContext::get().initialize_fast_dispatch(mesh_device.get());
-
-        auto free = internal::ServiceCoreManager::get().get_claimable_cores(device);
-        CoreCoord svc_core = free[free.size() / 2];
-
-        internal::ServiceCoreManager::get().claim(device, free);
-        DeviceAddr stop_addr = internal::ServiceCoreManager::get().allocate_l1(device, svc_core, sizeof(uint32_t));
-        DeviceAddr counter_addr = internal::ServiceCoreManager::get().allocate_l1(device, svc_core, sizeof(uint32_t));
-        DeviceAddr service_done_addr =
-            internal::ServiceCoreManager::get().allocate_l1(device, svc_core, sizeof(uint32_t));
-
-        experimental::DispatchContext::get().terminate_fast_dispatch(mesh_device.get());
-
-        // Initialise all three words to 0.
-        CoreCoord phys_core = device->worker_core_from_logical_core(svc_core);
-        tt_cxy_pair phys_cxy(device->id(), phys_core.x, phys_core.y);
-        const uint32_t zero = 0;
-        cluster.write_core(&zero, sizeof(uint32_t), phys_cxy, stop_addr);
-        cluster.write_core(&zero, sizeof(uint32_t), phys_cxy, counter_addr);
-        cluster.write_core(&zero, sizeof(uint32_t), phys_cxy, service_done_addr);
-
-        // Phase 2: launch persistent kernel (SD, non-blocking).
-        // The kernel handles FD dispatch go signals internally via notify_dispatch_core_done.
-        Program prog;
-        auto kernel = CreateKernel(
-            prog,
-            "tests/tt_metal/tt_metal/test_kernels/misc/service_counter.cpp",
-            svc_core,
-            DataMovementConfig{.processor = DataMovementProcessor::RISCV_0, .noc = NOC::RISCV_0_default});
-        SetRuntimeArgs(
-            prog, kernel, svc_core, {(uint32_t)stop_addr, (uint32_t)counter_addr, (uint32_t)service_done_addr});
-
-        tt::tt_metal::detail::CompileProgram(device, prog, /*force_slow_dispatch=*/true);
-        tt::tt_metal::detail::WriteRuntimeArgsToDevice(device, prog, /*force_slow_dispatch=*/true);
-        tt::tt_metal::detail::LaunchProgram(
-            device, prog, /*wait_until_cores_done=*/false, /*force_slow_dispatch=*/true);
-
-        auto read_counter = [&]() -> uint32_t {
-            uint32_t val = 0;
-            cluster.read_core(&val, sizeof(uint32_t), phys_cxy, counter_addr);
-            return val;
-        };
-
-        auto assert_incrementing = [&](const std::string& label) {
-            uint32_t a = read_counter();
-            std::this_thread::sleep_for(std::chrono::milliseconds(20));
-            uint32_t b = read_counter();
-            EXPECT_GT(b, a) << label << " (cycle " << cycle << "): counter stalled (" << a << " -> " << b << ")";
-        };
-
-        assert_incrementing("SD baseline");
-
-        // Phase 3: FD re-init
-        experimental::DispatchContext::get().initialize_fast_dispatch(mesh_device.get());
-        assert_incrementing("after FD init");
-
-        // Sharded L1 buffer
-        auto fd_buf = MeshBuffer::create(fd_l1_global, fd_l1_config, mesh_device.get());
-        std::vector<uint32_t> fd_src_vec(num_tiles * single_tile_size / sizeof(uint32_t));
-        std::iota(fd_src_vec.begin(), fd_src_vec.end(), 100);
-        EnqueueWriteMeshBuffer(mesh_device->mesh_command_queue(), fd_buf, fd_src_vec);
-        Finish(mesh_device->mesh_command_queue());
-
-        for (const auto& coord : MeshCoordinateRange(mesh_device->shape())) {
-            std::vector<uint32_t> dst;
-            ReadShard(mesh_device->mesh_command_queue(), dst, fd_buf, coord);
-            EXPECT_EQ(dst, fd_src_vec) << "Cycle " << cycle << ": sharded L1 readback failed in FD mode at " << coord;
-        }
-        assert_incrementing("FD steady-state");
-
-        experimental::DispatchContext::get().terminate_fast_dispatch(mesh_device.get());
-        assert_incrementing("after FD terminate");
-        CoreCoord grid = device->compute_with_storage_grid_size();
-        // Run random workloads to stress the dispatch path and dirty compute state.
-        auto programs = tt::tt_metal::distributed::test::utils::create_random_programs(num_programs, grid, 0);
-        for (uint32_t i = 0; i < num_programs; i++) {
-            auto random_workload = std::make_shared<MeshWorkload>();
-            random_workload->add_program(
-                MeshCoordinateRange(
-                    MeshCoordinate{0, 0}, MeshCoordinate{mesh_device->num_rows() - 1, mesh_device->num_cols() - 1}),
-                std::move(*programs[i]));
-            EnqueueMeshWorkload(mesh_device->mesh_command_queue(), *random_workload, true);
-        }
-        Finish(mesh_device->mesh_command_queue());
-
-        // Phase 4: stop kernel, wait via service_done_addr
-        const uint32_t one = 1;
-        cluster.write_core(&one, sizeof(uint32_t), phys_cxy, stop_addr);
-
-        constexpr int kTimeoutMs = 5000;
-        constexpr int kPollMs = 1;
-        for (int elapsed = 0; elapsed < kTimeoutMs; elapsed += kPollMs) {
-            uint32_t done = 0;
-            cluster.read_core(&done, sizeof(uint32_t), phys_cxy, service_done_addr);
-            if (done) {
-                break;
-            }
-            std::this_thread::sleep_for(std::chrono::milliseconds(kPollMs));
-            ASSERT_LT(elapsed + kPollMs, kTimeoutMs) << "Service kernel stop timed out on cycle " << cycle;
-        }
-
-        internal::ServiceCoreManager::get().release(device, free);
-    }
-}
-
-// Launch a persistent service kernel on a claimed FD-idle core while simultaneously
-// running standard FD workloads on the normal worker grid. Verifies both paths are
-// independent needed for prefill in Blitz.
-// The service counter keeps incrementing while FD programs dispatch.
-TEST_F(DispatchContextFixture, FDWorkloadAndServiceKernelConcurrent) {
-    const auto& rt_options = MetalContext::instance().rtoptions();
-    if (!rt_options.get_fast_dispatch()) {
-        GTEST_SKIP() << "This test requires Fast Dispatch mode.";
-    }
-    const auto& cluster = MetalContext::instance().get_cluster();
-    if (!cluster.is_ubb_galaxy() && cluster.arch() != tt::ARCH::BLACKHOLE) {
-        GTEST_SKIP() << "Service-core APIs require BH or UBB Galaxy.";
-    }
-
-    auto mesh_device = MeshDevice::create(MeshDeviceConfig(MeshShape(1, 1)));
-    IDevice* device = mesh_device->get_device(MeshCoordinate(0, 0));
-
-    auto claimable = internal::ServiceCoreManager::get().get_claimable_cores(device);
-    internal::ServiceCoreManager::get().claim(device, claimable);
-
-    // Allocate L1 communication words on each service core.
-    struct CoreAddrs {
-        DeviceAddr stop, counter, service_done;
-    };
-    std::unordered_map<CoreCoord, CoreAddrs> core_addrs;
-    for (const auto& core : claimable) {
-        core_addrs[core] = {
-            .stop = internal::ServiceCoreManager::get().allocate_l1(device, core, sizeof(uint32_t)),
-            .counter = internal::ServiceCoreManager::get().allocate_l1(device, core, sizeof(uint32_t)),
-            .service_done = internal::ServiceCoreManager::get().allocate_l1(device, core, sizeof(uint32_t)),
-        };
-        const uint32_t zero = 0;
-        CoreCoord phys = device->worker_core_from_logical_core(core);
-        tt_cxy_pair cxy(device->id(), phys.x, phys.y);
-        cluster.write_core(&zero, sizeof(uint32_t), cxy, core_addrs[core].stop);
-        cluster.write_core(&zero, sizeof(uint32_t), cxy, core_addrs[core].counter);
-        cluster.write_core(&zero, sizeof(uint32_t), cxy, core_addrs[core].service_done);
-    }
-
-    // Build one program targeting all service cores with per-core runtime args.
-    {
-        CoreRangeSet svc_cores(claimable);
-        Program svc_prog;
-        auto kernel = CreateKernel(
-            svc_prog,
-            "tests/tt_metal/tt_metal/test_kernels/misc/service_counter.cpp",
-            svc_cores,
-            DataMovementConfig{.processor = DataMovementProcessor::RISCV_0, .noc = NOC::RISCV_0_default});
-        for (const auto& core : claimable) {
-            SetRuntimeArgs(
-                svc_prog,
-                kernel,
-                core,
-                {(uint32_t)core_addrs[core].stop,
-                 (uint32_t)core_addrs[core].counter,
-                 (uint32_t)core_addrs[core].service_done});
-        }
-        auto svc_workload = std::make_shared<MeshWorkload>();
-        svc_workload->add_program(MeshCoordinateRange(MeshCoordinate{0, 0}, MeshCoordinate{0, 0}), std::move(svc_prog));
-        EnqueueMeshWorkload(mesh_device->mesh_command_queue(), *svc_workload, false);
-    }
-
-    // Verify all service cores are running by checking one representative counter.
-    const CoreCoord rep_core = claimable[0];
-    CoreCoord rep_phys = device->worker_core_from_logical_core(rep_core);
-    tt_cxy_pair rep_cxy(device->id(), rep_phys.x, rep_phys.y);
-
-    auto read_counter = [&]() -> uint32_t {
-        uint32_t val = 0;
-        cluster.read_core(&val, sizeof(uint32_t), rep_cxy, core_addrs[rep_core].counter);
-        return val;
-    };
-    auto assert_incrementing = [&](const std::string& label) {
-        uint32_t a = read_counter();
-        std::this_thread::sleep_for(std::chrono::milliseconds(20));
-        uint32_t b = read_counter();
-        EXPECT_GT(b, a) << label << ": counter stalled (" << a << " -> " << b << ")";
-    };
-
-    assert_incrementing("after service launch");
-
-    // Launch standard 5x5 FD workloads on the regular worker grid concurrently.
-    const uint32_t num_programs = 5;
-    auto programs = tt::tt_metal::distributed::test::utils::create_random_programs(num_programs, CoreCoord{5, 5}, 0);
-    for (uint32_t i = 0; i < num_programs; i++) {
-        auto workload = std::make_shared<MeshWorkload>();
-        workload->add_program(MeshCoordinateRange(MeshCoordinate{0, 0}, MeshCoordinate{0, 0}), std::move(*programs[i]));
-        EnqueueMeshWorkload(mesh_device->mesh_command_queue(), *workload, false);
-        assert_incrementing("during FD workload " + std::to_string(i));
-    }
-    Finish(mesh_device->mesh_command_queue());
-    assert_incrementing("after all FD workloads complete");
-
-    // Stop all service kernels and wait for clean exit on each.
-    const uint32_t one = 1;
-    for (const auto& core : claimable) {
-        CoreCoord phys = device->worker_core_from_logical_core(core);
-        tt_cxy_pair cxy(device->id(), phys.x, phys.y);
-        cluster.write_core(&one, sizeof(uint32_t), cxy, core_addrs[core].stop);
-    }
-    constexpr int kTimeoutMs = 5000;
-    constexpr int kPollMs = 1;
-    for (const auto& core : claimable) {
-        CoreCoord phys = device->worker_core_from_logical_core(core);
-        tt_cxy_pair cxy(device->id(), phys.x, phys.y);
-        for (int elapsed = 0; elapsed < kTimeoutMs; elapsed += kPollMs) {
-            uint32_t done = 0;
-            cluster.read_core(&done, sizeof(uint32_t), cxy, core_addrs[core].service_done);
-            if (done) {
-                break;
-            }
-            std::this_thread::sleep_for(std::chrono::milliseconds(kPollMs));
-            ASSERT_LT(elapsed + kPollMs, kTimeoutMs)
-                << "Service kernel stop timed out on core (" << core.x << "," << core.y << ")";
-        }
-    }
-
-    internal::ServiceCoreManager::get().release(device, claimable);
-}
-
-// Verify allocator correctness: alignment, non-overlap, bytes_available accounting, and deallocation.
-TEST_F(DispatchContextFixture, ServiceCoreAllocatorCorrectness) {
-    const auto& rt_options = MetalContext::instance().rtoptions();
-    if (!rt_options.get_fast_dispatch()) {
-        GTEST_SKIP() << "This test requires Fast Dispatch mode.";
-    }
-    const auto& cluster = MetalContext::instance().get_cluster();
-    if (!cluster.is_ubb_galaxy() && cluster.arch() != tt::ARCH::BLACKHOLE) {
-        GTEST_SKIP() << "Service-core APIs require BH or UBB Galaxy.";
-    }
-
-    auto mesh_device = MeshDevice::create(MeshDeviceConfig(MeshShape(1, 1)));
-    IDevice* device = mesh_device->get_device(MeshCoordinate(0, 0));
-
-    auto claimable = internal::ServiceCoreManager::get().get_claimable_cores(device);
-    CoreCoord core = claimable[0];
-    internal::ServiceCoreManager::get().claim(device, {core});
-
-    const DeviceAddr dram_align = MetalContext::instance().hal().get_alignment(HalMemType::DRAM);
-
-    // Allocate several buffers of varying sizes and verify properties.
-    const std::vector<size_t> sizes = {64, 128, 32, 256};
-    std::vector<DeviceAddr> addrs;
-    size_t initial_available = internal::ServiceCoreManager::get().bytes_available(device, core);
-
-    for (size_t sz : sizes) {
-        DeviceAddr addr = internal::ServiceCoreManager::get().allocate_l1(device, core, sz);
-        EXPECT_EQ(addr % dram_align, 0u) << "Allocation not DRAM-aligned";
-        addrs.push_back(addr);
-    }
-
-    // No pair of allocations should overlap.
-    for (size_t i = 0; i < addrs.size(); i++) {
-        for (size_t j = i + 1; j < addrs.size(); j++) {
-            DeviceAddr a_start = addrs[i], a_end = addrs[i] + sizes[i];
-            DeviceAddr b_start = addrs[j], b_end = addrs[j] + sizes[j];
-            EXPECT_TRUE(a_end <= b_start || b_end <= a_start) << "Allocations " << i << " and " << j << " overlap";
-        }
-    }
-
-    // bytes_available should have decreased.
-    size_t after_alloc = internal::ServiceCoreManager::get().bytes_available(device, core);
-    EXPECT_LT(after_alloc, initial_available);
-
-    // Deallocate one allocation and verify space is recovered.
-    internal::ServiceCoreManager::get().deallocate_l1(device, core, addrs[1]);
-    size_t after_dealloc = internal::ServiceCoreManager::get().bytes_available(device, core);
-    EXPECT_GT(after_dealloc, after_alloc);
-
-    // OOM: requesting more than the full L1 space must TT_FATAL.
-    constexpr size_t k1p5MB = 1536 * 1024;
-    EXPECT_THROW(internal::ServiceCoreManager::get().allocate_l1(device, core, k1p5MB), std::runtime_error);
-
-    internal::ServiceCoreManager::get().release(device, {core});
-}
-
-// Verify TT_FATAL fires on every documented misuse path:
-//   - get_claimable_cores / claim before FD is active
-//   - allocate_l1 / deallocate_l1 / bytes_available on an unclaimed core
-TEST_F(DispatchContextFixture, ServiceCoreFatalGuards) {
-    if (MetalContext::instance().rtoptions().get_fast_dispatch()) {
-        GTEST_SKIP() << "This test requires Slow Dispatch mode.";
-    }
-    const auto& cluster = MetalContext::instance().get_cluster();
-    if (!cluster.is_ubb_galaxy() && cluster.arch() != tt::ARCH::BLACKHOLE) {
-        GTEST_SKIP() << "Service-core APIs require BH or UBB Galaxy.";
-    }
-
-    auto mesh_device = MeshDevice::create(MeshDeviceConfig(MeshShape(1, 1)));
-    IDevice* device = mesh_device->get_device(MeshCoordinate(0, 0));
-
-    // Before FD is active, both query and claim must fatal.
-    EXPECT_THROW(internal::ServiceCoreManager::get().get_claimable_cores(device), std::exception);
-    EXPECT_THROW(internal::ServiceCoreManager::get().claim(device, {{0, 0}}), std::exception);
-
-    // Bring FD up and claim one core so we can test the unclaimed-core guards.
-    experimental::DispatchContext::get().initialize_fast_dispatch(mesh_device.get());
-    auto free = internal::ServiceCoreManager::get().get_claimable_cores(device);
-    CoreCoord svc_core = free[0];
-    internal::ServiceCoreManager::get().claim(device, {svc_core});
-
-    // {0,0} is a worker-grid core — never a service core — so all three guards must fatal on it.
-    CoreCoord unclaimed{0, 0};
-    EXPECT_THROW(internal::ServiceCoreManager::get().allocate_l1(device, unclaimed, 4), std::exception);
-    EXPECT_THROW(internal::ServiceCoreManager::get().deallocate_l1(device, unclaimed, 0), std::exception);
-    EXPECT_THROW(internal::ServiceCoreManager::get().bytes_available(device, unclaimed), std::exception);
-
-    internal::ServiceCoreManager::get().release(device, {svc_core});
-    experimental::DispatchContext::get().terminate_fast_dispatch(mesh_device.get());
 }
 
 }  // namespace tt::tt_metal::distributed::test
